tabletserver endtoend: regroup tests
Moved config tests to a different file.
